chore(registry): merge category and extract scripts into single script
