+++ data/registry/cpp.json
@@ -1,8 +1,8 @@
 {
   "metadata": {
-    "generated_at": "2026-03-13T11:13:54.320492+00:00",
+    "generated_at": "2026-04-03T12:52:45.291002+00:00",
     "language": "cpp",
-    "total_registry_entries": 2341,
+    "total_registry_entries": 2383,
     "matched_entries": 0
   },
   "configurations": []
